chats

+++ app/Http/Mainstay/Chat/ChatMainstayController.php
@@ -4,7 +4,6 @@ use App\Models\CoreEngine\LogicModels\Chat\ChatLogic;
 use App\Models\CoreEngine\LogicModels\Chat\ChatMessageLogic;
 use App\Models\CoreEngine\LogicModels\User\UserLogic;
 use App\Models\System\ControllersModel\MainstayController;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
@@ -30,8 +29,8 @@ class ChatMainstayController extends MainstayController
                 [DB::raw("CONCAT_WS(' ', last_name, first_name, middle_name) as name")]))
                 ->getOne()['name'];
         }
-        $chat = (new ChatLogic($data))->getOne();
-        if (empty($chat)) {
+        $chat = (new ChatLogic($data))->setJoin(['ChatUser'])->getChatAndUndeleteIt();
+        if (!$chat) {
             $chat = $this->createChat($data);
         }
         return response()->json($chat);
@@ -64,8 +63,8 @@ class ChatMainstayController extends MainstayController
             'page' => 'required',
             'pageSize' => 'required'
         ]);
-        $select = ['*', DB::raw("DATE_FORMAT(created_at, '%H:%i') as time")];
-        return response()->json((new ChatMessageLogic($data, $select))->order('desc', 'id')->getMessageList());
+        $select = ['*', DB::raw("DATE_FORMAT(created_at, '%H:%i') as time"),];
+        return response()->json((new ChatMessageLogic($data, $select))->setJoin(['Sender'])->order('desc', 'id')->getMessageList());
     }
 
     public function actionSendMessage($param = []) {
@@ -81,8 +80,12 @@ class ChatMainstayController extends MainstayController
     public function actionDeleteMessage($param = []) {
         $this->params = empty($param) ? $this->params : $param;
         $data = Validator::validate($this->params, [
-            'id' => 'required|integer|exists:chat_message,id'
+            'id' => 'required|integer|exists:chat_message,id',
+            'user_id' => 'required|integer|exists:user_entity,id'
         ]);
+        if ($data['user_id'] !== (string) auth()->id()) {
+            return false;
+        }
         return response()->json((new ChatMessageLogic())->deleteForeva($data['id']));
     }
 
@@ -98,7 +101,7 @@ class ChatMainstayController extends MainstayController
     public function actionChatDelete($param = []) {
         $this->params = empty($param) ? $this->params : $param;
         $data = Validator::validate($this->params, [
-            'id' => 'required|integer|exists:chat,id'
+            'chat_id' => 'required|integer|exists:chat,id'
         ]);
         $data['user_id'] = \auth()->id();
         return response()->json((new ChatLogic())->deleteChat($data));
